requirements.txt
pillow==12.1.0
flask==3.0.3
